requirements.txt
streamlit==1.24.0
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
opencv-python==4.8.0.74
numpy==1.24.3
streamlit-webrtc